clean up tests, organize by database requirements, enable database-specific tests via environment variable
