+++ package.json
@@ -18,6 +18,9 @@
     "package-extension": "npm run build && npm run -w malloy-vscode package-extension",
     "lint": "eslint '**/*.ts{,x}'",
     "test": "jest --runInBand",
+    "test-bigquery": "DIALECT=bigquery jest --runInBand",
+    "test-postgres": "DIALECT=postgres jest --runInBand",
+    "test-duckdb": "DIALECT=duckdb jest --runInBand",
     "docs-clean": "rm -rf docs/.jekyll-cache/ && rm -rf docs/_includes/generated/ && rm -rf docs/documentation && rm -rf js/generated/",
     "docs-prebuild": "npm run build && npm run docs-clean",
     "docs-build": "npm run docs-prebuild && ts-node --project tsconfig.docs.json --esm ./docs/_scripts/build_docs/index.ts && cd docs && bundle exec jekyll build",
