+++ packages/malloy-db-test/src/dialects/bigquery/malloy_query.spec.ts
@@ -12,24 +12,20 @@
  */
 
 import { test } from "@jest/globals";
-import { testModel } from "./models/faa_model";
-import { fStringEq } from "./test_utils";
+import { testModel } from "../../models/faa_model";
+import { describeIfDatabaseAvailable, fStringEq } from "../../test_utils";
 
 import * as malloy from "@malloydata/malloy";
 
 import { Query } from "@malloydata/malloy";
-import { BigQueryTestConnection, RuntimeList } from "./runtimes";
+import { BigQueryTestConnection, RuntimeList } from "../../runtimes";
 
 const runtimeList = new RuntimeList(["bigquery"]);
 const runtime = runtimeList.runtimeMap.get("bigquery");
 if (runtime === undefined) {
   throw new Error("Couldn't build runtime");
 }
-const bq = new BigQueryTestConnection(
-  "bigquery",
-  {},
-  { defaultProject: "malloy-data" }
-);
+const bq = runtime.connection as BigQueryTestConnection;
 
 function compileQueryFromQueryDef(
   model: malloy.ModelMaterializer,
@@ -56,7 +52,9 @@ async function bqCompile(sql: string): Promise<boolean> {
   return true;
 }
 
-describe("expression tests", () => {
+const [describe] = describeIfDatabaseAvailable(["bigquery"]);
+
+describe("BigQuery expression tests", () => {
   const faa = runtime._loadModelFromModelDef(testModel);
 
   it("simple_pipeline", async () => {
