+++ test/README.md
@@ -1,6 +1,10 @@
 ## Tests
 
-Running all tests requires access to both a BigQuery and a Postgres database. Below is information about setting up postgres:
+By default, tests run against BigQuery, Postgres, and DuckDB.
+
+Tests can also be run against a specific database, using the DATABASES environment variable, such as: `DATABASES=x,y npm run test`
+
+Setting up postgres:
 
 # Setup for Postgres Test Data
 
@@ -16,3 +20,10 @@ copies the test data in `malloytest-postgres.sql.gz` into the database.
 **postgres_stop.sh** - stops the postgres server
 
 **state_fact.sql** - example file on how to insert data from json
+
+Setting up DuckDB:
+
+# Building test database for duckdb
+
+1. At top-level, run `npx ts-node scripts/build_duckdb_test_database.ts`
+2. A file called `duckdb_test.db` should be created in the test/data/duckdb folder - tests will automatically look there.
